tweak(client/launcher): don't rebuild subprocesses for main-only files
